fix : 기존의 레디스 모듈 설정 클래스 제거 및 상품서비스 레디스 설정 클래스추가
