--- module-redis/src/main/java/com/github/msafriends/moduleredis/config/RedisConfig.java
@@ -1,32 +1,0 @@
-package com.github.msafriends.moduleredis.config;
-
-import org.springframework.cache.CacheManager;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-import org.springframework.stereotype.Component;
-
-import java.time.Duration;
-
-@Component
-public class RedisConfig {
-
-    public static RedisConnectionFactory createRedisConnectionFactory(String host, int port) {
-        return new LettuceConnectionFactory(host, port);
-    }
-
-    public static CacheManager createCacheManager(RedisConnectionFactory redisConnectionFactory) {
-        RedisCacheManager.RedisCacheManagerBuilder builder = RedisCacheManager.RedisCacheManagerBuilder.fromConnectionFactory(redisConnectionFactory);
-        RedisCacheConfiguration configuration = RedisCacheConfiguration.defaultCacheConfig()
-                .serializeKeysWith(RedisSerializationContext.SerializationPair.fromSerializer(new StringRedisSerializer()))
-                .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(new GenericJackson2JsonRedisSerializer())) // Value Serializer 변경
-                .prefixCacheNameWith("RedisCache:")
-                .entryTtl(Duration.ofMinutes(30));
-        builder.cacheDefaults(configuration);
-        return builder.build();
-    }
-}
